Change the way translationTable is used
It used to be the table, but it's now (since some months ago) in CakePHP
it is a registry alias. Neatly hidden in fixing that is discovering that
the hasOne and hasMany were attempting to share the same association
alias, (before it was association_translations for the hasOne and
AssociationTranslations for the hasMany), which means that just changing
how the translation table reference was found would cause
findTranslations to stop working.

So, config changed to account for the above