Fix bug and added tests
As explained in the documentation of the ```ShadowTranslateBehavior::_addFieldsToQuery()``` function:

> If the query is using autofields (directly or implicitly) add the main table's fields to the query first.

Checking if autoFields is being used (it's possible to both have specified fields in the select clause AND autoFields set to true) is never done, so eg. in a contained BTM association where the junctionTable fields are added to the query, the primary tables fields are never checked, the translation is never done on the BTM association.

This commit fixes that, by checking if the autoFields is true or false in the addFields function.